pyproject.toml
[tool.poetry]
name = "file-uploader-skeleton"
version = "0.1.0"
description = "Skeleton Project in FastAPI to upload file to S3 and fire some background Async jobs"
authors = ["Kuldeep Pisda <hello@kdpisda.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.26.0"
gunicorn = "^21.2.0"
sqlalchemy = "^2.0.25"
pymysql = "^1.1.0"
alembic = "^1.13.1"
cryptography = "^41.0.7"
fastapi-sqlalchemy = "^0.2.1"
celery = "^5.3.6"
python-multipart = "^0.0.6"
boto3 = "^1.34.20"
redis = "^5.0.1"
bcrypt = "^4.1.2"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
